refactor(node): extract chain subcommands to commands/chain.rs
Five cmd_chain_* functions move out of main.rs:
  cmd_chain_info, _validate, _block, _reset_trie, _verify_deep

Same pattern as #641 (wallet) + #643 (validator): pure CLI handlers,
no consensus path touched. verify-deep's local imports
(std::sync::Arc, sentrix::core::trie::{account_value_bytes,
address_to_key}) stay inside the function body — one-shot use, kept
local so the module header stays focused on what every command
shares.

CLI dispatch updated: `cmd_chain_*` → `commands::chain::cmd_chain_*`.

main.rs: 4,344 → 4,102 LOC (−242 = −5.6%).
Combined #641 + #643 + this: 4,780 → 4,102 (−678 = −14.2%).
