+++ bin/sentrix/src/commands/mod.rs
@@ -7,5 +7,6 @@
 //! validator key load, which the orchestration layer needs to drive
 //! directly. Everything else moves here, one logical group per file.
 
+pub mod chain;
 pub mod validator;
 pub mod wallet;
